Let language server subprocess be its own process group
In situations where `multilspy` is launched from another process, a
SIGINT or SIGTERM signal will be received by both the launching process
as well as the process running the language server because
`asyncio.create_subprocess_shell` defaults to `start_new_session=False`.
In such a situation, `lsp_protocol_handler.server.shutdown` will be
called but will hang since the underlying process is already terminated,
`lsp_protocol_handler.server` won't ever receive an answer back for
`shutdown`.

This change makes it so that SIGINT and SIGTERM won't propagate to the
language server process. We rely on the clean up code already in place
to shut down the process as desired.

See
[python documentation](https://docs.python.org/3.13/library/subprocess.html#subprocess.Popen)
for more details on `start_new_session`